requirements.txt
django
djangorestframework
psycopg2 - binary
Pillow
drf-yasg
python-decouple
djangorestframework-simplejwt
celery
django-filter